services/my-watchdog/requirements.txt
watchdog==4.0.0
pyyaml==6.0.1
requests==2.31.0
# Optional for more advanced actions (uncomment if needed)
# ffmpeg-python==0.2.0
# python-dotenv==1.0.1